refactor: migrate trpcClient to shared/api with stub
