+++ common_knowledge/Commonwealth-Frontend-Refactor-Plan.md
@@ -28,7 +28,7 @@ Legend: [ ] Not started, [~] In progress, [x] Done. Add a completion date in par
 - [~] 2.2 Create shared/ directory scaffolding + tsconfig/vite aliases — scoped `shared/hooks/*`, `shared/utils/*`, `shared/api/*`, and `features/*` mappings + `client/scripts/shared/*` scaffolding are in progress via issue #13412 while preserving existing `packages/commonwealth/shared/*` imports
 - [~] 2.3 Migrate shared reusable hooks to shared/hooks/ — audited 17-hook move to `client/scripts/shared/hooks/*` + legacy `client/scripts/hooks/*` re-export stubs in progress via issue #13413
 - [~] 2.4 Migrate shared utility helpers to shared/utils/ — shared helper/util module set migrated to `client/scripts/shared/utils/*` with legacy compatibility stubs in progress via issue #13414
-- [ ] 2.5 Migrate trpcClient to shared/api/
+- [~] 2.5 Migrate trpcClient to shared/api/ — source moved to `client/scripts/shared/api/trpcClient.ts` with legacy `utils/trpcClient.ts` compatibility re-export in progress via issue #13415
 - [ ] 2.6 Update consumer imports for shared hooks
 - [ ] 2.7 Update consumer imports for shared utils (batch 1)
 - [ ] 2.8 Update consumer imports for shared utils (batch 2)
